Refactor ClamAV scan execution to use 'clamdscan' in production environment for improved performance

+++ app/Controllers/File.php
@@ -162,16 +162,18 @@ class File extends Controller
 
         $filePath = $file->store('files', $fileName);
 
-        $clamdPath = 'clamdscan'; // Faster than clamscan
 
-        $scanResult = shell_exec(escapeshellarg($clamdPath) . " --no-summary " . escapeshellarg(WRITEPATH . 'uploads/' . $filePath));
+        if (ENVIRONMENT == 'production') {
+            $scanResult = shell_exec(escapeshellarg('clamdscan') . " --no-summary " . escapeshellarg(WRITEPATH . 'uploads/' . $filePath));
 
-        if (strpos($scanResult, 'OK') === false) {
-            if (file_exists($filePath)) {
-                unlink($filePath);
+            if (strpos($scanResult, 'OK') === false) {
+                if (file_exists($filePath)) {
+                    unlink($filePath);
+                }
+                return $this->response->setJSON(['status' => 'error', 'message' => 'Malicious file detected.'])->setStatusCode(400);
             }
-            return $this->response->setJSON(['status' => 'error', 'message' => 'Malicious file detected.'])->setStatusCode(400);
         }
+
 
         $uuid = service('uuid');
         $uuid4 = $uuid->uuid4()->toString();
